move inspect helpers into ray-independent package
